+++ tnoodle-server/src/main/kotlin/org/worldcubeassociation/tnoodle/server/crypto/StringEncryption.kt
@@ -1,4 +1,4 @@
-package org.worldcubeassociation.tnoodle.server.cryptography
+package org.worldcubeassociation.tnoodle.server.crypto
 
 import java.util.*
 import javax.crypto.Cipher
